requirements.txt
aiofiles==23.2.1
aiosmtplib==2.0.2
aiosqlite==0.17.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==3.2.2
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.5
MarkupSafe==2.1.5
orjson==3.10.0
packaging==24.0
passlib==1.7.4
pillow==10.3.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pypika-tortoise==0.1.6
pytest==8.1.1
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
tomli==2.0.1
tortoise-orm==0.20.0
typing_extensions==4.10.0
ujson==5.9.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
